requirements.txt
python-telegram-bot==21.9
groq==0.13.1
python-dotenv==1.0.1
aiohttp==3.9.1